api/app/services/ocr_service.py
import time
from typing import List, Optional
from api.app.models.schemas import ExtractedData, ShipmentHeader, Container, LayoutLine, BoundingBox
from api.app.core.validators import validator
import io
from PIL import Image
import numpy as np

# Lazy load Surya settings
surya_loaded = False
det_model = None
det_processor = None
rec_model = None
rec_processor = None

def load_surya():
    global surya_loaded, det_model, det_processor, rec_model, rec_processor
    if not surya_loaded:
        try:
            from surya.ocr import run_ocr
            from surya.model.detection.segformer import load_model as load_det_model, load_processor as load_det_processor
            from surya.model.recognition.model import load_model as load_rec_model
            from surya.model.recognition.processor import load_processor as load_rec_processor
            
            print("🧠 Loading Surya Models (forced CPU + Float32 for stability)...")
            import torch
            det_model = load_det_model(device="cpu", dtype=torch.float32)
            det_processor = load_det_processor()
            rec_model = load_rec_model(device="cpu", dtype=torch.float32)
            rec_processor = load_rec_processor()
            surya_loaded = True
            print("🚀 Surya Models Loaded (CPU+Float32 Mode)!")
        except ImportError:
            print("⚠️ Surya not installed.")

class OcrService:
    """
    Orchestrates the conversion of Documents -> Structured, Validated Data.
    """
    
    def process_document(self, file_contents: bytes, filename: str) -> ExtractedData:
        """
        Main entry point.
        1. OCR (Text/Layout Extraction) via Surya (or text extraction via Gemini directly)
        2. Entity Extraction via LLM (Gemini 1.5 Flash)
        3. Logic Validation (The Firewall)
        """
        start_time = time.time()
        
        # 1. Image Pre-processing / Loading
        # For MVP launch, we pass the bytes directly to Gemini Multimodal
        # Gemini 1.5 is excellent at reading text from images directly, often skipping the need for distinct OCR 
        # unless we need specific bounding boxes for UI highlighting.
        
        # For the "Split-Screen" highlighting to work, we ideally need bounding boxes. 
        # Implementing Surya here for that local precision.
        
        # NOTE: For this implementation, we will assume standard Gemini extraction first for speed/ease
        # and mock the bounding boxes or implement Surya in V2 if the user installs the heavy deps.
        
        try:
            print(f"🔄 [Process Document] Starting processing for: {filename}")
            
            # 1.1 Run Surya for Layout (Local M4)
            print("▶️ [Surya] Calling Surya OCR...")
            layout_lines = self._call_surya_ocr(file_contents)
            print(f"✅ [Surya] Finished. Found {len(layout_lines)} lines.")
            
            # 1.2 Run Gemini for Extraction (Cloud)
            print("▶️ [Gemini] Calling Gemini 1.5 Flash...")
            extracted_data = self._call_gemini_flash(file_contents)
            print("✅ [Gemini] Extraction successful.")
            
            # 1.3 Merge Layout into Result
            extracted_data.layout = layout_lines
            extracted_data.raw_text = "Extracted via Gemini 1.5 Flash + Surya OCR (Local)"
            print("✅ [Process Document] Merge complete.")
            
        except Exception as e:
            print(f"❌ [CRITICAL ERROR] Pipeline Failed: {e}")
            import traceback
            traceback.print_exc()
            print("⚠️ Falling back to MOCK data.")
            # Fallback to Mock if API fails
            raw_text = self._mock_ocr(file_contents)
            extracted_data = self._mock_llm_extraction(raw_text)

        # 3. Validation Step ("The Firewall")
        print("▶️ [Validation] Applying business rules...")
        validated_data = self._apply_validation_logic(extracted_data)
        
        validated_data.processing_time_ms = int((time.time() - start_time) * 1000)
        print(f"🏁 [Done] Processing complete in {validated_data.processing_time_ms}ms")
        
        return validated_data

    def _call_gemini_flash(self, file_contents: bytes) -> ExtractedData:
        """
        Calls Google Gemini 1.5 Flash with the document image.
        """
        import google.generativeai as genai
        from api.app.core.config import settings
        import json

        if not settings.GOOGLE_API_KEY:
            raise ValueError("GOOGLE_API_KEY not set")

        genai.configure(api_key=settings.GOOGLE_API_KEY)
        
        model = genai.GenerativeModel('models/gemini-flash-latest')
        
        # Prompt designed to return strict JSON matching our schema
        prompt = """
        You are a specialized Data Extraction Agent for Logistics.
        Extract the Bill of Lading data into the following JSON structure.
        Return ONLY the JSON. No markdown formatting.
        
        Schema:
        {
            "header": {
                "shipper": "string",
                "consignee": "string", 
                "notify_party": "string",
                "vessel_name": "string",
                "voyage_number": "string",
                "port_of_loading": "string",
                "port_of_discharge": "string",
                "scac_code": "string",
                "hbl_number": "string",
                "mbl_number": "string"
            },
            "containers": [
                {
                    "container_number": "string",
                    "seal_number": "string",
                    "package_count": 0,
                    "weight_gross": 0.0,
                    "volume_cbm": 0.0,
                    "description": "string"
                }
            ]
        }
        """
        
        # Pass the image data directly (supported by 1.5 Flash)
        response = model.generate_content([
            {'mime_type': 'application/pdf', 'data': file_contents},
            prompt
        ])
        
        # Clean response (remove ```json ... ```)
        text = response.text.replace('```json', '').replace('```', '').strip()
        data_dict = json.loads(text)
        
        # Parse into Pydantic Models
        header = ShipmentHeader(**data_dict.get("header", {}))
        containers = [Container(**c) for c in data_dict.get("containers", [])]
        
        return ExtractedData(header=header, containers=containers, confidence_score=1.0)
            
    def _call_surya_ocr(self, file_contents: bytes) -> list[LayoutLine]:
        """
        Runs Surya OCR locally to get text and bounding boxes.
        """
        try:
            load_surya()
            if not surya_loaded:
                return []
                
            from surya.ocr import run_ocr
            import pypdfium2 as pdfium
            
            images = []
            
            # 1. Try opening as Image (PNG/JPG)
            try:
                img = Image.open(io.BytesIO(file_contents)).convert("RGB")
                images.append(img)
            except Exception:
                # 2. If valid image fails, try PDF
                try:
                    pdf = pdfium.PdfDocument(file_contents)
                    for i in range(len(pdf)):
                        page = pdf[i]
                        # Render to PIL Image (scale=2 for better OCR resolution, typically 300dpi)
                        bitmap = page.render(scale=2) 
                        pil_image = bitmap.to_pil()
                        images.append(pil_image)
                except Exception as e:
                    print(f"⚠️ Could not load as Image or PDF: {e}")
                    return []

            if not images:
                print("⚠️ No images loaded from file.")
                return []
            
            # Run Inference
            # langs=["en"] is optional
            predictions = run_ocr(images, [["en"] * len(images)], det_model, det_processor, rec_model, rec_processor)
            
            # Process results (aggregating all pages)
            layout_lines = []
            
            # Predictions is a list of OCRResult objects (one per image/page)
            for page_idx, ocr_result in enumerate(predictions):
                # We need to know which page the line is on if we support multi-page highlighting
                # For now, let's assume single page or just append all
                
                img_w, img_h = images[page_idx].size
                
                for line in ocr_result.text_lines:
                    # line.bbox is [x1, y1, x2, y2]
                    bbox = line.bbox
                    
                    # Normalize to 0-1 range
                    x = bbox[0] / img_w
                    y = bbox[1] / img_h
                    w = (bbox[2] - bbox[0]) / img_w
                    h = (bbox[3] - bbox[1]) / img_h
                    
                    layout_box = BoundingBox(x=x, y=y, width=w, height=h)
                    # We might want to store page_number in LayoutLine in future
                    layout_lines.append(LayoutLine(text=line.text, bbox=layout_box))
                
            return layout_lines
            
        except Exception as e:
            print(f"⚠️ Surya Extraction Failed: {e}")
            import traceback
            traceback.print_exc()
            return []

    def _mock_ocr(self, file_contents: bytes) -> str:
        """
        TODO: Replace with actual Surya OCR call.
        """
        return "MOCK TEXT: Bill of Lading for Maersk Line. Container MSKU1234567..."

    def _mock_llm_extraction(self, raw_text: str) -> ExtractedData:
        """
        TODO: Replace with actual Google Gemini 1.5 Flash API call.
        Returns a roughly populated ExtractedData object.
        """
        # Mocking what the LLM might return
        header = ShipmentHeader(
            shipper="ACME Corp",
            consignee="Global Imports Ltd",
            scac_code="MAEU",
            pol_locode="CNSHA", # Shanghai
            pod_locode="USNYC"  # New York
        )
        
        container1 = Container(
            container_number="MSKU1234565", # Valid (Check digit 5)
            seal_number="123", 
            description="Electronics"
        )
        
        container2 = Container(
            container_number="MSKU1234568", # Invalid Checksum
            seal_number="124", 
            description="Widgets"
        )
        
        return ExtractedData(
            header=header,
            containers=[container1, container2],
            raw_text=raw_text,
            confidence_score=0.95
        )

    def _apply_validation_logic(self, data: ExtractedData) -> ExtractedData:
        """
        Applies the business rules from api.core.validators.
        Updates the model with validation flags.
        """
        # 1. Validate Header Data
        if data.header.scac_code:
            is_valid_scac = validator.validate_scac(data.header.scac_code)
            # You might want to flag this in the response if invalid, 
            # for now we assume the LLM tries its best or we strip it?
            # Let's keep it but maybe add a warning field to ShipmentHeader in the future.
            pass

        # 2. Validate Containers (Critical)
        for container in data.containers:
            is_valid = validator.validate_container_iso6346(container.container_number)
            container.is_valid_checksum = is_valid
            
            if not is_valid:
                container.validation_message = "Invalid ISO 6346 checksum."
        
        return data

ocr_service = OcrService()
